+++ Cargo.toml
@@ -3,10 +3,11 @@ resolver = "2"
 members = [
     "fuzz",
     "liana",
+    "lianad",
     "liana-gui",
     "liana-ui",
 ]
-default-members = ["liana", "liana-gui", "liana-ui"]
+default-members = ["liana", "lianad", "liana-gui", "liana-ui"]
 
 [patch.crates-io]
 iced_style = { git = "https://github.com/edouardparis/iced", branch = "patch-0.12.3"}
