+++ liana-gui/Cargo.toml
@@ -16,7 +16,8 @@ path = "src/main.rs"
 [dependencies]
 async-trait = "0.1"
 async-hwi = { version = "0.0.24" }
-liana = { path = "../liana", default-features = false, features = ["nonblocking_shutdown"] }
+liana = { path = "../liana" }
+lianad = { path = "../lianad", default-features = false, features = ["nonblocking_shutdown"] }
 liana-ui = { path = "../liana-ui" }
 backtrace = "0.3"
 hex = "0.4.3"
