+++ liana-gui/src/daemon/model.rs
@@ -3,17 +3,16 @@ use std::collections::{HashMap, HashSet};
 
 use liana::descriptors::LianaDescriptor;
 pub use liana::{
-    commands::{
-        CreateSpendResult, GetAddressResult, GetInfoResult, GetLabelsResult, LabelItem,
-        ListCoinsEntry, ListCoinsResult, ListSpendEntry, ListSpendResult, ListTransactionsResult,
-        TransactionInfo,
-    },
     descriptors::{LianaPolicy, PartialSpendInfo, PathSpendInfo},
     miniscript::bitcoin::{
         bip32::{DerivationPath, Fingerprint},
         psbt::Psbt,
         secp256k1, Address, Amount, Network, OutPoint, Transaction, Txid,
     },
+};
+pub use lianad::commands::{
+    CreateSpendResult, GetAddressResult, GetInfoResult, GetLabelsResult, LabelItem, ListCoinsEntry,
+    ListCoinsResult, ListSpendEntry, ListSpendResult, ListTransactionsResult, TransactionInfo,
 };
 
 pub type Coin = ListCoinsEntry;
