+++ lianad/src/bitcoin/d/utils.rs
