+++ lianad/src/bitcoin/electrum/utils.rs
